# backend-v2/src/routes/userSessions.ts
/*
 * ⚠️  WARNING: GENERATED CODE - DO NOT MODIFY ⚠️
 * 
 * This file is automatically generated by the backend code generator.
 * Any manual changes will be overwritten when the generator runs.
 * 
 * To make changes, modify the generator configuration files in:
 * /tools/backend-tools/backend-generator-v2/
 * 
 * Then re-run the generator to update this file.
 */

import { Router } from 'express';
import { authenticate } from '../middleware/auth';
import { validate } from '../middleware/validation';
import { UserSessionService } from '../services/user_sessionsService';
import { UserSessionResponse, CreateUserSessionRequest, UpdateUserSessionRequest } from '../models/UserSession';

const router = Router();
const user_sessionsService = new UserSessionService();

// POST /
router.post('/', authenticate, async (req: any, res) => {
  try {
    const result = await user_sessionsService.createUserSession(req.body);
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

// GET /:id
router.get('/:id', authenticate, async (req: any, res) => {
  try {
    const result = await user_sessionsService.getUserSessionById(req.params.id);
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

// PUT /:id
router.put('/:id', authenticate, async (req: any, res) => {
  try {
    const result = await user_sessionsService.updateUserSession(req.params.id, req.body);
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

// DELETE /:id
router.delete('/:id', authenticate, async (req: any, res) => {
  try {
    const result = await user_sessionsService.deleteUserSession(req.params.id);
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

// GET /
router.get('/', authenticate, async (req: any, res) => {
  try {
    const result = await user_sessionsService.listUserSessions();
    res.json({ success: true, data: result });
  } catch (error: any) {
    res.status(400).json({ success: false, error: error.message });
  }
});

export default router;